refactor: split build-helpers-internal lib.rs into focused submodules
