+++ crates/build-helpers-internal/src/hash.rs
@@ -1,0 +1,65 @@
+//! SHA-256 hashing and verification of downloaded artifacts.
+
+use std::io::Read;
+use std::path::Path;
+
+/// Computes the SHA-256 hash of a file using the `sha2` crate. Returns `None` on I/O error.
+fn sha256_file(path: &Path) -> Option<String> {
+    use sha2::{Digest, Sha256};
+
+    let mut file = match std::fs::File::open(path) {
+        Ok(f) => f,
+        Err(e) => {
+            println!(
+                "cargo:warning=Failed to open {:?} for SHA-256 verification: {}",
+                path, e
+            );
+            return None;
+        }
+    };
+
+    let mut hasher = Sha256::new();
+    let mut buffer = [0_u8; 8192];
+    loop {
+        let n = match file.read(&mut buffer) {
+            Ok(n) => n,
+            Err(e) => {
+                println!(
+                    "cargo:warning=Failed to read {:?} for SHA-256 verification: {}",
+                    path, e
+                );
+                return None;
+            }
+        };
+        if n == 0 {
+            break;
+        }
+        hasher.update(&buffer[..n]);
+    }
+
+    let hash = hasher.finalize();
+    let mut hex = String::with_capacity(hash.len() * 2);
+    for byte in hash {
+        use std::fmt::Write;
+        write!(hex, "{:02x}", byte).unwrap();
+    }
+    Some(hex)
+}
+
+/// Verifies the SHA-256 hash of a file against an expected value.
+/// Returns `true` if the hash matches.
+pub fn verify_sha256(path: &Path, expected: &str, label: &str) -> bool {
+    let Some(actual) = sha256_file(path) else {
+        return false;
+    };
+
+    if actual.eq_ignore_ascii_case(expected) {
+        true
+    } else {
+        println!(
+            "cargo:warning={} SHA-256 mismatch for {:?}: expected {}, got {}",
+            label, path, expected, actual
+        );
+        false
+    }
+}
